+++ worker/routes/webhooks/webhook-test-routes.ts
@@ -1,0 +1,121 @@
+import { Hono } from 'hono';
+import type { Env } from '../../core-utils';
+import { ok, bad, serverError } from '../../core-utils';
+import { logger } from '../../logger';
+import { CircuitBreaker } from '../../CircuitBreaker';
+import { RetryDelay } from '../../config/time';
+import type { Context } from 'hono';
+
+export function webhookTestRoutes(app: Hono<{ Bindings: Env }>) {
+  app.post('/api/webhooks/test', async (c: Context) => {
+    let body: { url: string; secret: string } = { url: '', secret: '' };
+
+    try {
+      body = await c.req.json<{
+        url: string;
+        secret: string;
+      }>();
+
+      if (!body.url || !body.secret) {
+        return bad(c, 'Missing required fields: url, secret');
+      }
+
+      const testPayload = {
+        id: `test-${crypto.randomUUID()}`,
+        eventType: 'test',
+        data: { message: 'Webhook test payload' },
+        timestamp: new Date().toISOString()
+      };
+
+      const encoder = new TextEncoder();
+      const key = encoder.encode(body.secret);
+      const data = encoder.encode(JSON.stringify(testPayload));
+
+      const importedKey = await crypto.subtle.importKey('raw', key, { name: 'HMAC', hash: 'SHA-256' }, false, ['sign']);
+      const signature = await crypto.subtle.sign('HMAC', importedKey, data);
+      const hash = new Uint8Array(signature);
+      const hexArray = Array.from(hash).map(b => b.toString(16).padStart(2, '0'));
+      const signatureHeader = `sha256=${hexArray.join('')}`;
+
+      const breaker = CircuitBreaker.createWebhookBreaker(body.url);
+
+      let lastError: Error | unknown;
+      const maxRetries = 3;
+      const retryDelaysMs = [RetryDelay.ONE_SECOND_MS, RetryDelay.TWO_SECONDS_MS, RetryDelay.THREE_SECONDS_MS];
+
+      for (let attempt = 0; attempt <= maxRetries; attempt++) {
+        try {
+          const response = await breaker.execute(async () => {
+            return await fetch(body.url, {
+              method: 'POST',
+              headers: {
+                'Content-Type': 'application/json',
+                'X-Webhook-Signature': signatureHeader,
+                'X-Webhook-ID': testPayload.id,
+                'X-Webhook-Timestamp': testPayload.timestamp,
+                'User-Agent': 'Akademia-Pro-Webhook/1.0'
+              },
+              body: JSON.stringify(testPayload),
+              signal: AbortSignal.timeout(30000)
+            });
+          });
+
+          const responseText = await response.text();
+
+          if (attempt > 0) {
+            logger.info('Webhook test succeeded after retry', {
+              url: body.url,
+              status: response.status,
+              attempt: attempt + 1,
+              retries: attempt
+            });
+          } else {
+            logger.info('Webhook test sent', { url: body.url, status: response.status });
+          }
+
+          return ok(c, {
+            success: response.ok,
+            status: response.status,
+            response: responseText
+          });
+        } catch (error) {
+          lastError = error;
+          const errorMessage = error instanceof Error ? error.message : 'Unknown error';
+
+          if (errorMessage.includes('Circuit breaker is open')) {
+            logger.warn('Webhook test skipped due to open circuit breaker', {
+              url: body?.url,
+              errorMessage
+            });
+            return ok(c, {
+              success: false,
+              error: 'Circuit breaker is open for this webhook URL. Please wait before retrying.'
+            });
+          }
+
+          if (attempt < maxRetries) {
+            const delay = retryDelaysMs[attempt];
+            logger.info('Webhook test retrying', {
+              url: body.url,
+              attempt: attempt + 1,
+              maxRetries: maxRetries + 1,
+              delayMs: delay
+            });
+            await new Promise(resolve => setTimeout(resolve, delay));
+          }
+        }
+      }
+
+      const finalErrorMessage = lastError instanceof Error ? lastError.message : 'Unknown error';
+      logger.error('Webhook test failed after all retries', finalErrorMessage);
+      return ok(c, {
+        success: false,
+        error: finalErrorMessage
+      });
+    } catch (error) {
+      const errorMessage = error instanceof Error ? error.message : 'Unknown error';
+      logger.error('Webhook test failed with error', errorMessage);
+      return serverError(c, 'Failed to test webhook');
+    }
+  });
+}
